requirements.txt
pydantic
duckduckgo-search
bs4
requests
rich
openai-agents
python-dotenv